Time: 5 ms (20.27%), Space: 40.3 MB (35.29%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean judgeSquareSum(int c){
         for(long i=0;i*i<=c;i++){
-            double val=Math.sqrt(c-(int)(i*i));
+            double val=Math.sqrt(c-i*i);
             if(val==(int)val) return true;
         }
         return false;
